deletiing unusing user data
