+++ src/entities/User.ts
@@ -1,18 +1,19 @@
-import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
+import { Entity, Column } from 'typeorm';
+// import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
 
 @Entity()
 export class User {
-  @PrimaryGeneratedColumn('uuid')
-  userId: string;
+  // @PrimaryGeneratedColumn('uuid')
+  // userId: string;
 
   @Column({ unique: true })
   email: string;
 
-  @Column({ unique: true })
-  firstName: string;
+  // @Column({ unique: true })
+  // firstName: string;
 
-  @Column({ unique: true })
-  lastName: string;
+  // @Column({ unique: true })
+  // lastName: string;
 
   @Column({ unique: true })
   passwordHash: string;
@@ -20,24 +21,23 @@ export class User {
   @Column({ default: false })
   verifiedEmail: boolean;
 
-  @Column({ unique: true })
-  money: number;
+  // @Column({ unique: true })
+  // money: number;
 
-  @Column({ unique: true })
-  profit: number;
+  // @Column({ unique: true })
+  // profit: number;
 
-  @Column({ unique: true })
-  loss: number;
+  // @Column({ unique: true })
+  // loss: number;
 }
 
-export class CryptoCurrency {
-  @Column({ unique: true })
-  type: string;
+// export class CryptoCurrency {
+//   @Column({ unique: true })
+//   type: string;
 
-  @Column({ unique: true })
-  value: number;
+//   @Column({ unique: true })
+//   value: number;
 
-  @Column({ unique: true })
-  preValue: number;
-}
-
+//   @Column({ unique: true })
+//   preValue: number;
+// }
