+++ src/models/UserModel.ts
@@ -3,17 +3,10 @@ import { User } from '../entities/User';
 
 const userRepository = AppDataSource.getRepository(User);
 
-async function addUser(
-  email: string,
-  firstName: string,
-  lastName: string,
-  passwordHash: string
-): Promise<User> {
+async function addUser(email: string, passwordHash: string): Promise<User> {
   // Create the new user object
   let newUser = new User();
   newUser.email = email;
-  newUser.firstName = firstName;
-  newUser.lastName = lastName;
   newUser.passwordHash = passwordHash;
 
   newUser = await userRepository.save(newUser);
@@ -26,9 +19,20 @@ async function getUserByEmail(email: string): Promise<User | null> {
   return user;
 }
 
-async function getUserById(userId: string): Promise<User | null> {
-  const user = await userRepository.findOne({ where: { userId } });
-  return user;
-}
+// async function getUserById(userId: string): Promise<User[] | null> {
+//   // const user = await userRepository.findOne({ where: { userId } });
+//   // return user;
 
-export { addUser, getUserByEmail, getUserById };
+//   if (!userId) {
+//     return null;
+//   }
+//   const user = await userRepository
+//     .createQueryBuilder('user')
+//     .where({ userId })
+//     .select(['user.email', 'user.userId'])
+//     .getMany();
+
+//   return user;
+// }
+
+export { addUser, getUserByEmail };
