+++ src/types/utils.d.ts
@@ -6,7 +6,7 @@ type DatabaseConstraintError = {
 
 type AuthRequest = {
   email: string;
-  firstName: string;
-  lastName: string;
+  // firstName: string;
+  // lastName: string;
   password: string;
 };
